WIP fixing extra additions to local storage
